currency_converter_backend/core/serializers/currency_serializer.py
from rest_framework import serializers
from core.models.currency import Currency

class CurrencySerializer(serializers.ModelSerializer):
    class Meta:
        model = Currency
        fields = ['id', 'code', 'name', 'symbol']
