currency_converter_backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.6.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
cfgv==3.4.0
channels==4.1.0
charset-normalizer==3.4.0
colorama==0.4.6
decorator==5.1.1
distlib==0.3.9
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
executing==2.1.0
filelock==3.16.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
identify==2.6.1
idna==3.10
ipython==8.28.0
jedi==0.19.1
matplotlib-inline==0.1.7
multidict==6.1.0
mysqlclient==2.2.4
nodeenv==1.9.1
parso==0.8.4
platformdirs==4.3.6
pre_commit==4.0.1
prompt_toolkit==3.0.48
propcache==0.2.0
pure_eval==0.2.3
Pygments==2.18.0
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.6
wcwidth==0.2.13
yarl==1.15.1